hw4_11.05

+++ model/index.js
@@ -1,28 +1,34 @@
 const Contact = require('./schemas/contact');
 
-const listContacts = async () => {
-  const result = await Contact.find({});
+const listContacts = async (userId) => {
+  const result = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email -_id',
+  });
   return result;
 };
 
-const getContactById = async (id) => {
-  const result = await Contact.findOne({ _id: id });
+const getContactById = async (userId, id) => {
+  const result = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'email -_id',
+  });
   return result;
 };
 
-const addContact = async (body) => {
-  const result = await Contact.create(body);
+const addContact = async (userId, body) => {
+  const result = await Contact.create({ ...body, owner: userId });
   return result;
 };
 
-const removeContact = async (id) => {
-  const result = Contact.findByIdAndRemove({ _id: id });
+const removeContact = async (userId, id) => {
+  const result = Contact.findByIdAndRemove({ _id: id, owner: userId });
   return result;
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (userId, id, body) => {
   const result = Contact.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true }
   );
